Extract PermissionInterrogator class
